plus

+++ BigDecimalInt.h
@@ -1,17 +1,18 @@
 #ifndef P2_BIGDECIMALINT_H
 #define P2_BIGDECIMALINT_H
 #include<bits/stdc++.h>
+
 using namespace std;
 class BigDecimalInt{
 private:
     string num;
 
 public:
-    
-    BigDecimalInt(string decStr); 
+    // constructor
+    BigDecimalInt(string decStr); // rejects bad inputs
 
 
-    BigDecimalInt(int decInt);
+    BigDecimalInt(long long int decInt);
     string getstr();
 
     BigDecimalInt operator+(BigDecimalInt anotherDec);
@@ -24,7 +25,7 @@ public:
 
     bool operator==(BigDecimalInt anotherDec);
 
-    BigDecimalInt operator=(BigDecimalInt anotherDec); 
+    BigDecimalInt operator=(BigDecimalInt anotherDec); // lsa msh 3arfa
 
     int size();
